Fix bug with volume slider
In recent versions of Chrome (80+) and Firefox (74) there is a bug in
the player: if the player once was playing, adjusting the volume opens
a new tab.

This happens because jQuery("a").attr("target", "_blank"); adds a target
to the volume slider.

This PR excludes the a.mejs-horizontal-volume-slider from adding a
target attribute, which fixes the annoying new tab bug.

Tested in Chromium 80 and Firefox 74.